+++ src/request_list.js
@@ -8,6 +8,7 @@ import Request from './request';
 import events from './events';
 import { ACTOR_EVENT_NAMES } from './constants';
 import { getFirstKey } from './utils';
+import { getValue, setValue } from './key_value_store';
 
 // TODO: better tests
 const URL_REGEX = '(http|https)://[\\w-]+(\\.[\\w-]+)+([\\w-.,@?^=%&:/~+#-]*[\\w@?^=%&;/~+#-])?';
@@ -45,16 +46,9 @@ const ensureUniqueKeyValid = (uniqueKey) => {
  *         // Note that all URLs must start with http:// or https://
  *         { requestsFromUrl: 'http://www.example.com/my-url-list.txt', userData: { isFromUrl: true } },
  *     ],
- *     // Initialize from previous state if act was restarted due to some error
- *     state: await Apify.getValue('my-request-list-state'),
  * });
  *
  * await requestList.initialize(); // Load requests.
- *
- * // Save state of the RequestList instance every 5 seconds.
- * setInterval(() => {
- *      Apify.setValue('my-request-list-state', requestList.getState());
- * }, 5000);
  *
  * // Get requests from list
  * const request1 = requestList.fetchNextRequest();
@@ -89,6 +83,9 @@ const ensureUniqueKeyValid = (uniqueKey) => {
  *     },
  * }
  * ```
+ * @param {String} [options.persistStateKey] Key-value store key under which RequestList persists its state. If this is set then RequestList
+ *                                           persists its state in regular intervals and loads the state from there in a case thats restarted
+ *                                           due to some error or migration to another worker machine.
  */
 export default class RequestList {
     constructor(opts = {}) {
@@ -126,7 +123,7 @@ export default class RequestList {
 
         // If this key is set then we persist url list into default key-value store under this key.
         this.persistStateKey = persistStateKey;
-        this.isCurrentStatePersisted = true;
+        this.isStatePersisted = true;
 
         this.isLoading = false;
         this.isInitialized = false;
@@ -186,9 +183,12 @@ export default class RequestList {
                 if (!this.persistStateKey) return;
 
                 events.on(ACTOR_EVENT_NAMES.PERSIST_STATE, () => {
-                    if (this.isCurrentStatePersisted) return;
+                    if (this.isStatePersisted) return;
 
                     return setValue(this.persistStateKey, this.getState())
+                        .then(() => {
+                            this.isStatePersisted = true;
+                        })
                         .catch((err) => {
                             log.exception(err, 'Cannot persist state of UrlList', { persistStateKey: this.persistStateKey });
                         });
@@ -268,7 +268,7 @@ export default class RequestList {
                     const request = this.requests[this.nextIndex];
                     this.inProgress[request.uniqueKey] = true;
                     this.nextIndex++;
-                    this.isCurrentStatePersisted = false;
+                    this.isStatePersisted = false;
                     return request;
                 }
 
@@ -294,7 +294,7 @@ export default class RequestList {
                 this._ensureIsInitialized();
 
                 delete this.inProgress[uniqueKey];
-                this.isCurrentStatePersisted = false;
+                this.isStatePersisted = false;
             });
     }